simulator/requirements.txt
gmqtt==0.6.14
uvloop==0.19.0; sys_platform != 'win32'
python-dotenv
